requirements.txt
# SPDX-FileCopyrightText: 2024 NewYork-Presbyterian Hospital
#
# SPDX-License-Identifier: MIT

amqp==5.2.0
asgiref>=3.4.1
attrs==23.2.0
beautifulsoup4==4.10.0
billiard==4.2.0
black==24.4.0
celery==5.3.6
certifi==2021.5.30
cfgv==3.4.0
charset-normalizer==2.0.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Deprecated==1.2.14
distlib==0.3.8
Django==3.2.13
django-appconf==1.0.4
django-auth-ldap==3.0.0
django-compressor==2.4.1
django-cors-headers==3.11.0
django-csp==3.7
django-env==0.2.1
django-environ==0.8.1
django-libsass==0.9
django-rq==2.10.2
djangorestframework==3.15.1
exceptiongroup==1.2.0
filelock==3.14.0
flake8==7.0.0
identify==2.5.36
idna==3.2
iniconfig==2.0.0
isort==5.13.2
kombu==5.3.6
libsass==0.21.0
lxml==4.9.2
mando==0.7.1
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.21.3
packaging==24.0
pandas==1.3.4
pathspec==0.12.1
pdfkit==1.0.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit
prompt-toolkit==3.0.43
py==1.11.0
pyaes==1.6.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==6.2.5
pytest-pythonpath==0.7.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-ldap==3.4.4
pytz==2021.1
PyYAML==6.0.1
radon==6.0.1
rcssmin==1.0.6
reportlab==3.6.5
requests==2.26.0
rjsmin==1.1.0
rq==1.16.1
six==1.16.0
soupsieve==2.3
SQLAlchemy==2.0.29
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.7
vine==5.1.0
virtualenv==20.26.2
wcwidth==0.2.13
wrapt==1.16.0